+++ src/core/Button/ButtonOutlined/index.ts
@@ -1,2 +1,1 @@
-export * from './Default/ButtonOutlined';
-export * from './Highlight/ButtonOutlined';
+export * from './ButtonOutlined';
