+++ src/core/Button/ButtonOutlined/styles.ts
@@ -1,6 +1,7 @@
-import { makeStyles } from '@material-ui/core';
+import { fade, makeStyles, Theme } from '@material-ui/core';
 
-const useGlobalStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme: Theme) => ({
+  // Global
   buttonOutline: {
     minWidth: '6.875rem',
     height: '2.8125rem',
@@ -11,6 +12,26 @@ const useGlobalStyles = makeStyles(() => ({
   valueField: {
     fontSize: '0.75rem',
   },
+
+  // Button Outlined
+  border: {
+    borderColor: theme.palette.highlight,
+    '&:hover': {
+      borderColor: theme.palette.highlight,
+      background: 'transparent',
+      boxShadow: `${fade(theme.palette.highlight, 0.5)} 0 0.3rem 0.4rem 0`,
+    },
+    '&:disabled': {
+      borderColor: theme.palette.disabledBackground,
+    },
+  },
+  disabled: {
+    color: theme.palette.text.disabled,
+  },
+  text: {
+    color: (props) =>
+      props !== true ? theme.palette.highlight : theme.palette.text.primary,
+  },
 }));
 
-export default useGlobalStyles;
+export default useStyles;
